added threshold and system working on uart

+++ Project_4/Knock_detector/Includes/accelerometer.h
@@ -86,7 +86,7 @@
 ******************************************************************************/
 
 unsigned char AccData[6];
-float Xout_14_bit, Yout_14_bit, Zout_14_bit;
+uint16_t Xout_14_bit, Yout_14_bit, Zout_14_bit;
 char Xout_14_bit_in_ascii, Yout_14_bit_in_ascii, Zout_14_bit_in_ascii;
 float Xout_g, Yout_g, Zout_g;
 volatile char DataReady;
